requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
multidict==6.1.0
openai==0.28.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.16
requests==2.32.3
sniffio==1.3.1
starlette==0.40.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.17.1